+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   - Support Auto refresh settings memoizing with a ttl of 1 second avoiding the need of restarting server when changing configs. #502
   - Bump clj-kondo adding new `gen-class` linter and other fixes/improvements. Fixes #589
   - Remove unused duplicate require if any. #527 
+  - Fix crash on clean-ns when ns contains comments.
   
 - Editor
   - Add reference code lens for ns forms. #578